feat: move search hint into service
